requirements.txt
moviepy
PyQt5
argparse
colour
numpy
Pillow
progressbar
scipy
tqdm
opencv-contrib-python-headless
pycairo
pydub
pygments
pyreadline; sys_platform == 'win32'
PyQt5
matplotlib
torch